Revert "Revert "Dokarkivproxy del av dokarkiv & trenger ikke oms-proxy (#103)" (#104)"
This reverts commit 2c4ad848c282e334ad74175e944fabb31216516b.
