+++ README.md
@@ -36,6 +36,6 @@ Interne henvendelser kan sendes via Slack i kanalen #sif_omsorgspenger.
 
 Dokumentation på integrasjoner i bruk:<br>
 <a href="https://confluence.adeo.no/display/BOA/Utviklerdokumentasjon+-+start+her">Joark</a> slack: #team_dokumentløsinger<br>
-<a href="https://confluence.adeo.no/display/TO/Systemdokumentasjon+Oppgave#">Oppgave</a> slack: #mfn<br>
+<a href="https://confluence.adeo.no/display/TO/Systemdokumentasjon+Oppgave#">Oppgave</a> slack: #team-oppgavehåndtering<br>
 <a href="https://confluence.adeo.no/display/BOA/saf#">Saf</a> slack: #saf
