+++ nais/prod-gcp.json
@@ -5,18 +5,17 @@
   ],
   "externalHosts": [
     "dokarkiv.prod-fss-pub.nais.io",
-    "omsorgspenger-proxy.prod-fss-pub.nais.io"
+    "oppgave.prod-fss-pub.nais.io",
+    "saf.prod-fss-pub.nais.io"
   ],
   "kafkaPool": "nav-prod",
   "azureTenant": "nav.no",
   "env": {
     "DOKARKIV_BASE_URL": "https://dokarkiv.prod-fss-pub.nais.io",
     "DOKARKIV_SCOPES": "api://prod-fss.teamdokumenthandtering.dokarkiv/.default",
-    "OPPGAVE_BASE_URL": "https://omsorgspenger-proxy.prod-fss-pub.nais.io/oppgave",
-    "OPPGAVE_SCOPES": "e816bb42-bf62-4be9-a1cc-a9be70fe4403/.default",
-    "DOKARKIVPROXY_BASE_URL": "https://omsorgspenger-proxy.prod-fss-pub.nais.io/dokarkivproxy",
-    "DOKARKIVPROXY_SCOPES": "e816bb42-bf62-4be9-a1cc-a9be70fe4403/.default",
-    "SAF_BASE_URL": "https://omsorgspenger-proxy.prod-fss-pub.nais.io/saf",
-    "SAF_SCOPES": "e816bb42-bf62-4be9-a1cc-a9be70fe4403/.default"
+    "OPPGAVE_BASE_URL": "https://oppgave.prod-fss-pub.nais.io",
+    "OPPGAVE_SCOPES": "api://prod-fss.oppgavehandtering.oppgave/.default",
+    "SAF_BASE_URL": "https://saf.prod-fss-pub.nais.io",
+    "SAF_SCOPES": "api://prod-fss.teamdokumenthandtering.saf/.default"
   }
 }
