+++ src/main/kotlin/no/nav/omsorgspenger/joark/DokarkivClient.kt
@@ -10,8 +10,12 @@ import no.nav.helse.dusseldorf.ktor.client.SimpleHttpClient.readTextOrThrow
 import no.nav.helse.dusseldorf.oauth2.client.AccessTokenClient
 import no.nav.omsorgspenger.AzureAwareClient
 import no.nav.omsorgspenger.CorrelationId
+import no.nav.omsorgspenger.Fagsystem
+import no.nav.omsorgspenger.Identitetsnummer
 import no.nav.omsorgspenger.JournalpostId
 import no.nav.omsorgspenger.JournalpostId.Companion.somJournalpostId
+import no.nav.omsorgspenger.Saksnummer
+import org.intellij.lang.annotations.Language
 import org.json.JSONObject
 import java.net.URI
 
@@ -30,6 +34,8 @@ internal class DokarkivClient(
     private fun JournalpostId.oppdaterJournalpostUrl() = "$baseUrl/rest/journalpostapi/v1/journalpost/${this}"
     private fun JournalpostId.ferdigstillJournalpostUrl() =
         "$baseUrl/rest/journalpostapi/v1/journalpost/${this}/ferdigstill"
+    private fun knyttTilAnnenSakUrl(journalpostId: JournalpostId) =
+        "$baseUrl/rest/journalpostapi/v1/journalpost/$journalpostId/knyttTilAnnenSak"
 
     internal suspend fun opprettJournalpost(
         correlationId: CorrelationId,
@@ -80,6 +86,44 @@ internal class DokarkivClient(
         check(httpStatus.isSuccess()) {
             "Feil ved ferdigstilling av journalpost. HttpStatus=[${httpStatus.value}, Response=[$responseBody], Url=[$url]"
         }
+    }
+
+    internal suspend fun knyttTilAnnenSak(
+        journalpostId: JournalpostId,
+        fagsystem: Fagsystem,
+        identitetsnummer: Identitetsnummer,
+        saksnummer: Saksnummer,
+        correlationId: CorrelationId
+    ): JournalpostId {
+        @Language("JSON")
+        val dto = """
+        {
+            "sakstype": "FAGSAK",
+            "fagsaksystem": "${fagsystem.name}",
+            "fagsakId": "$saksnummer",
+            "journalfoerendeEnhet": "9999",
+            "tema": "OMS",
+            "bruker": {
+                "idType": "FNR",
+                "id": "$identitetsnummer"
+            }
+        }
+        """.trimIndent()
+
+        val url = knyttTilAnnenSakUrl(journalpostId = journalpostId)
+        val (httpStatusCode, response) = url.httpPut { builder ->
+            builder.header("Nav-CallId", "$correlationId")
+            builder.header("Nav-Consumer-Id", "omsorgspenger-journalforing")
+            builder.accept(ContentType.Application.Json)
+            builder.header(HttpHeaders.Authorization, authorizationHeader())
+            builder.jsonBody(dto)
+        }.readTextOrThrow()
+
+        require(httpStatusCode.isSuccess()) {
+            "Feil fra Dokarkivproxy. URL=[$url], HttpStatusCode=[${httpStatusCode.value}], Response=[$response]"
+        }
+
+        return JSONObject(response).getString("nyJournalpostId").somJournalpostId()
     }
 
     private fun HttpRequestBuilder.defaultHeaders(correlationId: CorrelationId) {
